+++ qiskit_nature/second_q/mappers/__init__.py
@@ -25,6 +25,7 @@ operators.
    :nosignatures:
 
    QubitMapper
+   ModeBasedMapper
 
 FermionicOp Mappers
 +++++++++++++++++++
@@ -101,6 +102,7 @@ from .bosonic_linear_mapper import BosonicLinearMapper
 from .logarithmic_mapper import LogarithmicMapper
 from .direct_mapper import DirectMapper
 from .qubit_mapper import QubitMapper
+from .mode_based_mapper import ModeBasedMapper
 from .interleaved_qubit_mapper import InterleavedQubitMapper
 from .tapered_qubit_mapper import TaperedQubitMapper
 
@@ -116,4 +118,5 @@ __all__ = [
     "QubitMapper",
     "InterleavedQubitMapper",
     "TaperedQubitMapper",
+    "ModeBasedMapper",
 ]
